requirements.txt
# requirements.txt
# Required packages for Research Assistant Agentic Chatbot
langchain
huggingface_hub
python-dotenv
requests
certifi
